# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FixMyMind API",
        "description": "API for therapeutic voice and text chat with AI",
        "contact": {
            "email": "contact@fixmymind.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.fixmymind.org/api",
            "description": "Production API Server"
        },
        {
            "url": "/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/chat": {
            "post": {
                "tags": [
                    "Text Chat"
                ],
                "summary": "Send a text message to the AI therapist",
                "operationId": "4d0be0e9abf1d0625c11b162939065ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "messages": {
                                        "description": "Array of conversation messages",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "role": {
                                                    "type": "string",
                                                    "enum": [
                                                        "user",
                                                        "assistant"
                                                    ]
                                                },
                                                "content": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AI response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reply": {
                                            "description": "AI therapist response",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/voice-chat": {
            "post": {
                "tags": [
                    "Voice Chat"
                ],
                "summary": "Handle voice chat upload and response",
                "operationId": "72dc15d9cd052af6fcf3745096b1fb7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Audio file to upload (supported formats: wav, mp3, mp4, m4a, flac, webm, ogg)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transcript": {
                                            "description": "Transcribed text from audio",
                                            "type": "string"
                                        },
                                        "text": {
                                            "description": "AI response text",
                                            "type": "string"
                                        },
                                        "audio_url": {
                                            "description": "URL to generated audio response",
                                            "type": "string",
                                            "format": "url"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/voice-sessions/{id}/messages": {
            "post": {
                "tags": [
                    "Voice Sessions"
                ],
                "summary": "Store a new message in a voice session",
                "operationId": "5851497d054a3d7d4e60c416dfb064ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Audio file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transcript": {
                                            "type": "string"
                                        },
                                        "text": {
                                            "type": "string"
                                        },
                                        "audio_url": {
                                            "type": "string"
                                        },
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/voice-sessions/{id}": {
            "get": {
                "tags": [
                    "Voice Sessions"
                ],
                "summary": "Get all messages for a voice session",
                "operationId": "af6f8d4b391e9a19467eae394160bc65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "session": {
                                            "type": "integer"
                                        },
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "sender": {
                                                        "type": "string"
                                                    },
                                                    "text": {
                                                        "type": "string"
                                                    },
                                                    "audio_url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "timestamp": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}